Move ControlCollectionEnumerator to a separate file
